[hand-control] rename mapping_control to hand_control for better understanding
