+++ aerial_robot_planning/scripts/hand_control/sub_pos_objects.py
